Added macro migration rules as well
