+++ clang/test/dpct/pytorch/c10.cu
@@ -1,15 +1,12 @@
 // RUN: rm -rf %T/pytorch/c10
 // RUN: mkdir -p %T/pytorch/c10/src
 // RUN: cp %S/c10.cu %T/pytorch/c10/src/
-// RUN: cp %S/user_defined_rule_pytorch.yaml %T/pytorch/c10/
-// RUN: cp -r %S/pytorch_cuda_inc %T/pytorch/c10/
+// RUN: cp -r %S/pytorch_inc %T/pytorch/c10/
 // RUN: cd %T/pytorch/c10
 // RUN: mkdir dpct_out
-// RUN: dpct -out-root dpct_out %T/pytorch/c10/src/c10.cu --extra-arg="-I%T/pytorch/c10/pytorch_cuda_inc" --cuda-include-path="%cuda-path/include" --rule-file=%T/pytorch/c10/user_defined_rule_pytorch.yaml  -- -x cuda --cuda-host-only
+// RUN: dpct -out-root dpct_out %T/pytorch/c10/src/c10.cu --extra-arg="-I%T/pytorch/c10/pytorch_inc" --cuda-include-path="%cuda-path/include" --rule-file=%S/../../../tools/dpct/DpctOptRules/pytorch_api.yaml  -- -x cuda --cuda-host-only
 // RUN: FileCheck --input-file %T/pytorch/c10/dpct_out/c10.dp.cpp --match-full-lines %T/pytorch/c10/src/c10.cu
-// RUN: %if build_lit %{icpx -c -fsycl -DNO_BUILD_TEST  %T/pytorch/c10/dpct_out/c10.dp.cpp -o %T/pytorch/c10/dpct_out/c10.dp.o %}
 
-#ifndef NO_BUILD_TEST
 #include <iostream>
 // CHECK: #include <c10/core/DeviceGuard.h>
 #include <c10/cuda/CUDAGuard.h>
@@ -17,6 +14,19 @@
 #include <c10/cuda/CUDAStream.h>
 // CHECK: #include <c10/xpu/XPUFunctions.h>
 #include <c10/cuda/CUDAFunctions.h>
+// CHECK: #include <c10/xpu/XPUMacros.h>
+#include <c10/cuda/CUDAMacros.h>
+
+// CHECK: #define BE_BUILD_SHARED_LIBS C10_XPU_BUILD_SHARED_LIBS
+// CHECK-NEXT: #define BE_EXPORT C10_XPU_EXPORT
+// CHECK-NEXT: #define BE_IMPORT C10_XPU_IMPORT
+// CHECK-NEXT: #define BE_API C10_XPU_API
+// CHECK-NEXT: #define BE_BUILD_MAIN_LIB C10_XPU_BUILD_MAIN_LIB
+#define BE_BUILD_SHARED_LIBS C10_CUDA_BUILD_SHARED_LIBS
+#define BE_EXPORT C10_CUDA_EXPORT
+#define BE_IMPORT C10_CUDA_IMPORT
+#define BE_API C10_CUDA_API
+#define BE_BUILD_MAIN_LIB C10_CUDA_BUILD_MAIN_LIB
 
 int main() {
   // device APIs
@@ -53,9 +63,11 @@ int main() {
   // CHECK: auto currentStream = c10::xpu::getCurrentXPUStream();
   auto currentStream = c10::cuda::getCurrentCUDAStream();
 
-  // CHECK: dpct::queue_ptr curr_cuda_st = &(currentStream.queue());
-  // CHECK-NEXT: curr_cuda_st = &(c10::xpu::getCurrentXPUStream().queue());
+  // CHECK: dpct::queue_ptr curr_cuda_st =
+  // CHECK-NEXT: &static_cast<sycl::queue &>(currentStream.queue());
   cudaStream_t curr_cuda_st = currentStream.stream();
+  // CHECK: curr_cuda_st =
+  // CHECK-NEXT: &static_cast<sycl::queue &>(c10::xpu::getCurrentXPUStream().queue());
   curr_cuda_st = c10::cuda::getCurrentCUDAStream().stream();
 
   // CHECK: auto deviceStream = c10::xpu::getCurrentXPUStream(0);
@@ -63,5 +75,3 @@ int main() {
 
   return 0;
 }
-
-#endif
