--- clang/test/dpct/pytorch/pytorch_cuda_inc/ATen/cuda/CUDAContext.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <ATen/cuda/CUDAContextLight.h>
-
-#include <ATen/Context.h>
-#include <ATen/cuda/Exceptions.h>
-#include <c10/cuda/CUDAStream.h>
-#include <c10/util/Logging.h>
