+++ clang/test/dpct/pytorch/pytorch_inc/ATen/Tensor.h
@@ -1,0 +1,1 @@
+#pragma once
