+++ clang/test/dpct/pytorch/pytorch_inc/ATen/core/Tensor.h
@@ -1,0 +1,1 @@
+#pragma once
