+++ clang/test/dpct/pytorch/pytorch_inc/ATen/cuda/CUDAContext.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <c10/cuda/CUDAStream.h>
+
+namespace at {
+using namespace c10;
+}
