+++ clang/test/dpct/pytorch/pytorch_inc/ATen/cuda/CUDATensorMethods.cuh
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <ATen/Tensor.h>
+#include <c10/util/Half.h>
+
+#include <cuda.h>
+#include <cuda_runtime.h>
+#include <cuda_fp16.h>
