+++ clang/test/dpct/pytorch/pytorch_inc/c10/cuda/CUDAMacros.h
@@ -1,0 +1,5 @@
+#define C10_CUDA_BUILD_SHARED_LIBS
+#define C10_CUDA_EXPORT
+#define C10_CUDA_IMPORT
+#define C10_CUDA_API
+#define C10_CUDA_BUILD_MAIN_LIB
