+++ clang/test/dpct/pytorch/pytorch_inc/c10/util/Half.h
@@ -1,0 +1,1 @@
+#pragma once
