+++ clang/test/dpct/pytorch/pytorch_inc/torch/torch.h
@@ -1,0 +1,6 @@
+namespace torch {
+class Tensor {
+    public:
+        bool is_cuda();
+};
+}
