--- clang/test/dpct/pytorch/user_defined_rule_pytorch.yaml
@@ -1,92 +1,0 @@
----
-# ATen header rules
-- Rule: rule_ATen_core_Tensor_h
-  Kind: Header
-  Priority: Takeover
-  In: ATen/core/Tensor.h
-  Out: "<ATen/core/Tensor.h>"
-
-- Rule: rule_ATen_cuda_CUDAContext_h
-  Kind: Header
-  Priority: Takeover
-  In: ATen/cuda/CUDAContext.h
-  Out: "<ATen/xpu/XPUContext.h>"
-
-# stream rules
-- Rule: rule_c10_cuda_CUDAStream_h
-  Kind: Header
-  Priority: Takeover
-  In: c10/cuda/CUDAStream.h
-  Out: "<c10/xpu/XPUStream.h>"
-
-- Rule: rule_c10_cuda_CUDAStream
-  Kind: Class
-  Priority: Takeover
-  In: c10::cuda::CUDAStream
-  Out: c10::xpu::XPUStream
-  Methods:
-    - In: stream
-      Out: "&($method_base queue())"
-
-- Rule: rule_c10_cuda_getCurrentCUDAStream
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::getCurrentCUDAStream
-  Out: c10::xpu::getCurrentXPUStream($1)
-
-# device rules
-- Rule: rule_c10_cuda_CUDAGuard_h
-  Kind: Header
-  Priority: Takeover
-  In: c10/cuda/CUDAGuard.h
-  Out: "<c10/core/DeviceGuard.h>"
-
-- Rule: rule_c10_cuda_OptionalCUDAGuard
-  Kind: Type
-  Priority: Takeover
-  In: c10::cuda::OptionalCUDAGuard
-  Out: c10::OptionalDeviceGuard
-  Includes: ["<c10/core/DeviceGuard.h>"]
-
-# CUDAFunctions rules
-- Rule: rule_c10_cuda_CUDAFunctions_h
-  Kind: Header
-  Priority: Takeover
-  In: c10/cuda/CUDAFunctions.h
-  Out: "<c10/xpu/XPUFunctions.h>"
-
-- Rule: rule_c10_cuda_device_count
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::device_count
-  Out: c10::xpu::device_count()
-
-- Rule: rule_c10_cuda_device_count_ensure_non_zero
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::device_count_ensure_non_zero
-  Out: c10::xpu::device_count_ensure_non_zero()
-
-- Rule: rule_c10_cuda_current_device
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::current_device
-  Out: c10::xpu::current_device()
-
-- Rule: rule_c10_cuda_set_device
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::set_device
-  Out: c10::xpu::set_device($1)
-
-- Rule: rule_c10_cuda_exchange_device
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::ExchangeDevice
-  Out: c10::xpu::exchange_device($1)
-
-- Rule: rule_c10_cuda_maybe_exchange_device
-  Kind: API
-  Priority: Takeover
-  In: c10::cuda::MaybeExchangeDevice
-  Out: c10::xpu::maybe_exchange_device($1)
