--- clang/test/dpct/pytorch_api_type/CUDAContext.h
@@ -1,1 +1,0 @@
-/*empty file*/
